# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Nunu Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
        "/article/getArticleListByCID": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "分类获取公开文章列表",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetArticleListByCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ArticleList"
                        }
                    }
                }
            }
        },
...
        "/v1/article/createArticle": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "新建文章",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CreateArticleResponseData"
                        }
                    }
                }
            }
        },
        "/v1/article/deleteArticle": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteArticleResponseData"
                        }
                    }
                }
            }
        },
        "/v1/article/deleteArticleList": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "批量删除文章",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DelArticleListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteArticleResponseData"
                        }
                    }
                }
            }
        },
        "/v1/article/getArticle": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "获取文章详细",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ArticleData"
                        }
                    }
                }
            }
        },
        "/v1/article/getArticleListByEs": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "es文章查询",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetArticleListByEsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.SearchArticleResp"
                        }
                    }
                }
            }
        },
        "/v1/article/getUserArticleList": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "获取个人文章列表",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetUserArticleListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ArticleList"
                        }
                    }
                }
            }
        },
        "/v1/article/updateArticle": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文章模块"
                ],
                "summary": "修改文章内容",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ArticleData"
                        }
                    }
                }
            }
        },
...
        "/v1/user/getCollege": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "获取学院信息",
                "parameters": [
                    {
                        "description": "params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetCollegeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CollegeResponseData"
                        }
                    }
                }
            }
        },
        "/v1/user/getCollegeList": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "获取学院列表",
                "responses": {
                    "200": {
                        "description": "返回学院信息列表",
                        "schema": {
                            "$ref": "#/definitions/v1.GetCollegeListDataResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1.ArticleData": {
            "type": "object",
            "properties": {
                "article_id": {
                    "description": "文章id",
                    "type": "integer"
                },
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "category": {
                    "description": "文章分类",
                    "type": "string"
                },
                "category_id": {
                    "description": "文章分类ID",
                    "type": "integer"
                },
                "comment_disabled": {
                    "description": "是否禁用评论",
                    "type": "boolean"
                },
                "content": {
                    "description": "文章内容",
                    "type": "string"
                },
                "content_short": {
                    "description": "文章摘要",
                    "type": "string"
                },
                "created_at": {
                    "description": "文章创建时间",
                    "type": "string"
                },
                "importance": {
                    "description": "文章重要性",
                    "type": "integer"
                },
                "kb_id": {
                    "description": "知识库ID",
                    "type": "integer"
                },
                "kb_name": {
                    "description": "知识库名称",
                    "type": "string"
                },
                "source_uri": {
                    "description": "文章外链",
                    "type": "string"
                },
                "status": {
                    "description": "文章状态",
                    "type": "integer"
                },
                "title": {
                    "description": "文章标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "文章更新时间",
                    "type": "string"
                },
                "uploaded_files": {
                    "description": "上传的文件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.FileUpload"
                    }
                }
            }
        },
        "v1.ArticleList": {
            "type": "object",
            "properties": {
                "article_data_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.ArticleData"
                    }
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
...
        "v1.CollegeResponseData": {
            "type": "object",
            "properties": {
                "collegeId": {
                    "type": "integer"
                },
                "collegeName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "v1.CreateArticleRequest": {
            "type": "object",
            "required": [
                "author_id",
                "content",
                "status",
                "title"
            ],
            "properties": {
                "author_id": {
                    "description": "作者ID",
                    "type": "string"
                },
                "category_id": {
                    "description": "文章分类ID",
                    "type": "integer"
                },
                "comment_disabled": {
                    "description": "是否禁用评论",
                    "type": "boolean"
                },
                "content": {
                    "description": "文章内容",
                    "type": "string"
                },
                "content_short": {
                    "description": "文章摘要",
                    "type": "string"
                },
                "importance": {
                    "description": "文章重要性",
                    "type": "integer"
                },
                "kb_id": {
                    "description": "知识库ID",
                    "type": "integer"
                },
                "source_uri": {
                    "description": "文章外链",
                    "type": "string"
                },
                "status": {
                    "description": "文章状态 0: 草稿 1: 发布",
                    "type": "integer"
                },
                "title": {
                    "description": "文章标题",
                    "type": "string"
                },
                "uploaded_files": {
                    "description": "上传的文件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.FileUpload"
                    }
                }
            }
        },
        "v1.CreateArticleResponseData": {
            "type": "object",
            "properties": {
                "article_id": {
                    "description": "文章ID",
                    "type": "integer"
                }
            }
        },
...
        "v1.DelArticleListReq": {
            "type": "object",
            "properties": {
                "article_id_List": {
                    "description": "文章ID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "v1.DeleteArticleRequest": {
            "type": "object",
            "properties": {
                "article_id": {
                    "description": "文章ID",
                    "type": "integer"
                }
            }
        },
        "v1.DeleteArticleResponseData": {
            "type": "object",
            "properties": {
                "deleted_count": {
                    "description": "删除的文章数量",
                    "type": "integer"
                }
            }
        },
...
        "v1.FileUpload": {
            "type": "object",
            "properties": {
                "file_name": {
                    "description": "文件名",
                    "type": "string"
                },
                "file_url": {
                    "description": "文件URL",
                    "type": "string"
                }
            }
        },
        "v1.GetArticleListByCategoryReq": {
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "文章分类ID",
                    "type": "integer"
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                }
            }
        },
        "v1.GetArticleListByEsReq": {
            "type": "object",
            "properties": {
                "adv_search": {
                    "description": "是否启用高级搜索",
                    "type": "boolean"
                },
                "categories": {
                    "description": "分类id，用于筛选",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "column": {
                    "description": "排序字段，通常是 \"_score\"",
                    "type": "string"
                },
                "content": {
                    "description": "搜索的内容关键词",
                    "type": "string"
                },
                "create_time_end": {
                    "description": "文章结束时间",
                    "type": "string"
                },
                "create_time_start": {
                    "description": "文章创建时间",
                    "type": "string"
                },
                "importance": {
                    "description": "文章重要性",
                    "type": "string"
                },
                "kb_id": {
                    "description": "知识库ID",
                    "type": "integer"
                },
                "keywords": {
                    "description": "搜索的关键字数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "order": {
                    "description": "排序方式，\"asc\" 或 \"desc\"",
                    "type": "string"
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "phrase_match": {
                    "description": "是否启用短语匹配",
                    "type": "boolean"
                },
                "title": {
                    "description": "搜索的标题",
                    "type": "string"
                }
            }
        },
        "v1.GetCollegeListDataResponse": {
            "type": "object",
            "properties": {
                "collegeList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.CollegeResponseData"
                    }
                }
            }
        },
        "v1.GetCollegeRequest": {
            "type": "object",
            "properties": {
                "collegeId": {
                    "type": "integer"
                }
            }
        },
...
        "v1.GetUserArticleListReq": {
            "type": "object",
            "properties": {
                "category_id": {
                    "description": "文章分类ID",
                    "type": "integer"
                },
                "created_at": {
                    "description": "文章创建时间",
                    "type": "string"
                },
                "created_end": {
                    "description": "文章结束时间",
                    "type": "string"
                },
                "kb_id": {
                    "description": "知识库ID",
                    "type": "integer"
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "status": {
                    "description": "文章状态",
                    "type": "integer"
                },
                "title": {
                    "description": "文章标题",
                    "type": "string"
                }
            }
        },
...
        "v1.SearchArticleResp": {
            "type": "object",
            "properties": {
                "articles": {
                    "description": "文章列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.ArticleSearchInfo"
                    }
                },
                "page_index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total_count": {
                    "description": "总记录数",
                    "type": "integer"
                }
            }
        },
...
        "v1.UpdateArticleRequest": {
            "type": "object",
            "required": [
                "author_id",
                "content",
                "status",
                "title"
            ],
            "properties": {
                "article_id": {
                    "description": "文章ID",
                    "type": "integer"
                },
                "author_id": {
                    "description": "作者ID",
                    "type": "string"
                },
                "category_id": {
                    "description": "文章分类ID",
                    "type": "integer"
                },
                "comment_disabled": {
                    "description": "是否禁用评论",
                    "type": "boolean"
                },
                "content": {
                    "description": "文章内容",
                    "type": "string"
                },
                "content_short": {
                    "description": "文章摘要",
                    "type": "string"
                },
                "importance": {
                    "description": "文章重要性",
                    "type": "integer"
                },
                "kb_id": {
                    "description": "知识库ID",
                    "type": "integer"
                },
                "source_uri": {
                    "description": "文章外链",
                    "type": "string"
                },
                "status": {
                    "description": "文章状态 0: 草稿 1: 发布",
                    "type": "integer"
                },
                "title": {
                    "description": "文章标题",
                    "type": "string"
                },
                "uploaded_files": {
                    "description": "上传的文件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.FileUpload"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}